+++ test/utils.peer.js
@@ -14,7 +14,7 @@ const defaultServerAddrs = ['/ip4/127.0.0.1/tcp/5334/ws']
 const Rendezvous = require('../src')
 const Server = require('../src/server')
 
-const U = module.exports = {
+const Utils = module.exports = {
   createSwarm: (id, addrs, post) => new Promise((resolve, reject) => {
     Id.createFromJSON(id, (err, peerID) => {
       if (err) return reject(err)
@@ -42,17 +42,17 @@ const U = module.exports = {
     })
   }),
   createServer: async (id, conf, addrs) => {
-    const swarm = await U.createSwarm(id, addrs || defaultServerAddrs)
+    const swarm = await Utils.createSwarm(id, addrs || defaultServerAddrs)
     const server = new Server(Object.assign(Object.assign({}, conf || {}), {swarm}))
     server.start()
     return server
   },
   createRendezvousPeer: async (id, conf, addrs) => {
     let rendezvous
-    const swarm = await U.createSwarm(id, addrs || defaultAddrs, (swarm) => {
+    const swarm = await Utils.createSwarm(id, addrs || defaultAddrs, (swarm) => {
       rendezvous = new Rendezvous(swarm, conf || {})
     })
-    await promisify(swarm.dial.bind(swarm, await U.createServerPeerInfo()))()
+    await promisify(swarm.dial.bind(swarm, await Utils.createServerPeerInfo()))()
     return rendezvous
   },
   createServerPeerInfo: async () => {
